REFACTOR: move event dto to kafka module
