+++ kafka/src/main/kotlin/vp/togedo/kafka/data/email/ValidCodeEventDto.kt
@@ -1,6 +1,7 @@
-package vp.togedo.data.email
+package vp.togedo.kafka.data.email
 
 data class ValidCodeEventDto(
     val code: String,
     val email: String
 )
+
