+++ notification/core/src/main/kotlin/vp/togedo/listener/EmailEventListener.kt
@@ -3,7 +3,7 @@ package vp.togedo.listener
 import com.fasterxml.jackson.databind.ObjectMapper
 import org.springframework.kafka.annotation.KafkaListener
 import org.springframework.stereotype.Component
-import vp.togedo.data.email.ValidCodeEventDto
+import vp.togedo.kafka.data.email.ValidCodeEventDto
 import vp.togedo.service.EmailService
 
 @Component
